Bug 1293613: Don't assume that all videos start at 0. r?gerald
The tests incorrectly assume that all videos start at 0. However this is often not the case, in particular for the mp4 files.
The buffered range is the intersection of the audio track and the video track. As such, if the video track starts at a later time than 0, the buffered range of the source buffer can't be starting at 0.

Rather than using different videos, we properly use the correct values; this is done to ensure that buffered range are calculated correctly, regardless of the video content.

timestamps verify with mkvinfo utility for webm and ffprobe for mp4.

See issue https://github.com/w3c/web-platform-tests/issues/1939

MozReview-Commit-ID: AMKgJHEJsWr